Do not attempt to manage home directories.
This resolves a bug in the `gemrc` provider when using `name :local` or
`gemrc "local"` in which the resource could potentially recursively
change file permissions for the home directory. We now skip the
`directory` resource call within the `gemrc` custom resource when the
gemrc resource name resolves to either `:local` or `"local"` to ensure
files and directories do not become tainted when the user or group
values are set.

Additionally, this cleans up the chefspec slightly and adds a context
to make it a little cleaner.

Signed-off-by: Ryan Hass <rhass@chef.io>